Move IOChannel into its own header
It felt rather strange for it to be embedded in one of its subclass's
headers.
