+++ include/oxen/quic/stream.hpp
@@ -10,8 +10,9 @@
 #include <variant>
 #include <vector>
 
-#include "datagram.hpp"
+#include "connection_ids.hpp"
 #include "error.hpp"
+#include "iochannel.hpp"
 #include "types.hpp"
 #include "utils.hpp"
